pyproject.toml
[tool.poetry]
name = "softdesksupport"
version = "0.1.0"
description = "Soft Desk Support API"
authors = ["JasonBats <46994866+JasonBats@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
Django = "5.0.4"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
PyJWT = "2.8.0"
sqlparse = "0.5.0"
tzdata = "2024.1"
freezegun = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
